Huge code cleanup. Snake case for files, separated kernels into different files, comments, ... Created special LZ77 for 1 element per cycle Created a simple CRC32 function to validate output data
