--- DirectProgramming/DPC++FPGA/ReferenceDesigns/gunzip/src/Decompressor.hpp
@@ -1,1193 +1,0 @@
-#ifndef __DECOMPRESSOR_HPP__
-#define __DECOMPRESSOR_HPP__
-
-#include <CL/sycl.hpp>
-#include <sycl/ext/intel/fpga_extensions.hpp>
-#include <CL/sycl/INTEL/ac_types/ac_int.hpp>
-
-#include "ByteBitStream.hpp"
-#include "HeaderData.hpp"
-#include "unrolled_loop.hpp"
-#include "tuple.hpp"
-
-using namespace sycl;
-
-// we only use unsigned ac_ints, so use this alias to avoid having to type
-// 'false' all the time
-template<int bits>
-using ac_uint = ac_int<bits, false>;
-
-#ifdef BIT_BUFFER_BITS
-constexpr int kBitBufferBits = BIT_BUFFER_BITS;
-#else
-constexpr int kBitBufferBits = 48;
-#endif
-constexpr int kBitBufferMaxReadBits = 5;
-constexpr int kBitBufferMaxShiftBits = 30;
-using BitStreamT =
-  ByteBitStream<kBitBufferBits, kBitBufferMaxReadBits, kBitBufferMaxShiftBits>;
-
-//
-// Append a flag to a type 'T'
-//
-template<typename T>
-struct FlagBundle {
-  FlagBundle() : data(T()), flag(false) {}
-  FlagBundle(T d_in) : data(d_in), flag(false) {}
-  FlagBundle(T d_in, bool f_in) : data(d_in), flag(f_in) {}
-  FlagBundle(bool f_in) : data(T()), flag(f_in) {}
-
-  T data;
-  bool flag;
-};
-
-//
-// The data that comes out of the huffman decoder
-//
-struct HuffmanData {
-  HuffmanData() : len_or_sym(0), dist_or_flag(0) {}
-  HuffmanData(short symbol) : len_or_sym(symbol), dist_or_flag(-1) {}
-  HuffmanData(short len_in, short dist_in) :
-    len_or_sym(len_in), dist_or_flag(dist_in) {}
-  short len_or_sym;
-  short dist_or_flag; 
-};
-
-//
-// Holds a set (pack) of literals
-//
-template<int n>
-struct LiteralPack {
-  static constexpr int count_bits = fpga_tools::Log2(n) + 1;
-  unsigned char literal[n];
-  ac_uint<count_bits> valid_count;
-};
-
-//
-// Returns the number of trailing zero bits of the input
-//
-template<int bits>
-auto ctz(const ac_uint<bits>& in) {
-  constexpr int out_bits = fpga_tools::Log2(bits) + 1;
-  //ac_uint<out_bits> ret(bits);
-  ac_uint<out_bits> ret;
-  #pragma unroll
-  for (int i = bits - 1; i >= 0; i--) {
-    if (in[i]) {
-      ret = i;
-    }
-  }
-  return ret;
-}
-
-class HeaderKernelID;
-class HuffmanDecoderKernelID;
-class LZ77DecoderKernelID;
-class LiteralStackerKernelID;
-
-class HeaderToHuffmanPipeID;
-class HuffmanToLZ77PipeID;
-class LZ77ToLiteralStackerPipeID;
-
-template<typename InPipe, typename OutPipe, unsigned literals_per_cycle>
-event SubmitLiteralStackerKernel(queue& q) {
-  return q.single_task<LiteralStackerKernelID>([=] {
-    using OutPipeBundleT = FlagBundle<LiteralPack<literals_per_cycle>>;
-    constexpr int cache_idx_bits = fpga_tools::Log2(literals_per_cycle*2) + 1;
-
-    bool done;
-
-    // NOTE: cache_idx is fmax bottleneck of this kernel
-    // could shannonize to improve, but not worth it yet
-    ac_uint<cache_idx_bits> cache_idx = 0;
-    [[intel::fpga_register]] unsigned char cache_buf[literals_per_cycle * 2];
-
-    do {
-      bool data_valid;
-      auto pipe_data = InPipe::read(data_valid);
-      done = pipe_data.flag && data_valid;
-
-      if (data_valid && !done) {
-        #pragma unroll
-        for (int i = 0; i < literals_per_cycle; i++) {
-          if (i < pipe_data.data.valid_count) {
-            cache_buf[cache_idx + i] = pipe_data.data.literal[i];
-          }
-        }
-        cache_idx += pipe_data.data.valid_count;
-      }
-
-      if (cache_idx >= literals_per_cycle || done) {
-        // create the output pack of characters from the current cache
-        LiteralPack<literals_per_cycle> out_pack;
-        #pragma unroll
-        for (int i = 0; i < literals_per_cycle; i++) {
-          // copy the character
-          out_pack.literal[i] = cache_buf[i];
-
-          // shift the extra characters to the front of the cache
-          cache_buf[i] = cache_buf[i + literals_per_cycle];
-        }
-
-        // mark output with number of valid bytes
-        if (cache_idx <= literals_per_cycle) {
-          out_pack.valid_count = cache_idx;
-        } else {
-          out_pack.valid_count = literals_per_cycle;
-        }
-
-        // decrement cache_idx by number of elements we read
-        // it is safe to always subtract literals_per_cycle since that will only
-        // happen on the last iteration of the outer while loop (when 'done'
-        // is true) 
-        cache_idx -= literals_per_cycle;
-
-        // write output
-        OutPipe::write(OutPipeBundleT(out_pack));
-      }
-    } while (!done);
-    
-    // notify downstream kernel that we are done
-    OutPipe::write(OutPipeBundleT(true));
-  });
-}
-
-// TODO: support literals_per_cycle = 1
-template<typename InPipe, typename OutPipe, unsigned literals_per_cycle>
-event SubmitLZ77DecoderKernel(queue& q) {
-  return q.single_task<LZ77DecoderKernelID>([=] {
-    static_assert(literals_per_cycle > 0);
-    static_assert(fpga_tools::IsPow2(literals_per_cycle));
-    using OutPipeBundleT = FlagBundle<LiteralPack<literals_per_cycle>>;
-
-    bool done;
-    bool reading_history = false;
-    bool reading_history_next;
-    short history_counter;
-
-    constexpr unsigned kMaxHistory = 32768;
-    constexpr unsigned history_buffer_count = kMaxHistory / literals_per_cycle;
-    constexpr unsigned history_buffer_buffer_idx_bits = fpga_tools::Log2(literals_per_cycle);
-    constexpr unsigned history_buffer_buffer_idx_mask = literals_per_cycle - 1;
-    constexpr unsigned history_buffer_idx_bits = fpga_tools::Log2(history_buffer_count);
-    constexpr unsigned history_buffer_idx_mask = history_buffer_count - 1;
-    
-    using HistBufBufIdxT = ac_uint<history_buffer_buffer_idx_bits>;
-    using HistBufIdxT = ac_uint<history_buffer_idx_bits>;
-
-    HistBufBufIdxT history_buffer_buffer_idx = 0;
-    [[intel::fpga_register]] HistBufIdxT history_buffer_idx[literals_per_cycle];
-    fpga_tools::NTuple<unsigned char[history_buffer_count], literals_per_cycle> history_buffer;
-
-    constexpr int kCacheDepth = 7;
-    [[intel::fpga_register]] unsigned char history_buffer_cache_val[literals_per_cycle][kCacheDepth + 1];
-    [[intel::fpga_register]] HistBufIdxT history_buffer_cache_idx[literals_per_cycle][kCacheDepth + 1];
-
-    HistBufBufIdxT read_history_buffer_buffer_idx = 0;
-    [[intel::fpga_register]] HistBufIdxT read_history_buffer_idx[literals_per_cycle];
-    [[intel::fpga_register]] HistBufBufIdxT read_history_shuffle_idx[literals_per_cycle];
-
-    // precompute the function
-    //   dist_small + ((i - dist_small) % dist_small)
-    constexpr auto mod_lut_init = [&] {
-      constexpr int dim = literals_per_cycle - 1;
-      std::array<unsigned char, dim * dim> ret{};
-      for (int y = 0; y < dim; y++) {
-        for (int x = y; x < dim; x++) {
-          unsigned char dist = y + 1;
-          unsigned char i = x + 1;
-          ret[y * dim + x] = dist + ((i - dist) % dist);
-        }
-      }
-      return ret;
-    }();
-    
-    // ac_ints cannot be constexpr, so initialize an array of ac_int with
-    // the mod_lut_init ROM that was computed above
-    [[intel::fpga_register]] HistBufBufIdxT mod_lut[literals_per_cycle - 1][literals_per_cycle - 1];
-    for (int y = 0; y < (literals_per_cycle - 1); y++) {
-      for (int x = y; x < (literals_per_cycle - 1); x++) {
-        mod_lut[y][x] = mod_lut_init[y * (literals_per_cycle - 1) + x];
-      }
-    }
-
-    // initialize index pointers for each history buffer
-    #pragma unroll
-    for (int i = 0; i < literals_per_cycle; i++) { history_buffer_idx[i] = 0; }
-
-    [[intel::ivdep(kCacheDepth)]]
-    do {
-      bool data_valid = true;
-      LiteralPack<literals_per_cycle> out_data;
-
-      // if we aren't currently reading from the history, read from input pipe
-      if (!reading_history) {
-        // read from pipe
-        auto pipe_data = InPipe::read(data_valid);
-
-        // check if we are done
-        done = pipe_data.flag && data_valid;
-
-        // grab the symbol or the length and distance pair
-        short len_or_sym = pipe_data.data.len_or_sym;
-        short dist = pipe_data.data.dist_or_flag;
-
-        out_data.literal[0] = len_or_sym & 0xFF;
-        out_data.valid_count = 1;
-
-        // if we get a length distance pair we will read 'len_or_sym' bytes
-        // starting 'dist'
-        history_counter = len_or_sym;
-        reading_history = (dist > 0) && data_valid;
-        reading_history_next = history_counter > literals_per_cycle;
-
-        // grab the low Log2(literals_per_cycle) bits of the distance
-        HistBufBufIdxT dist_small = dist & history_buffer_buffer_idx_mask;
-        
-        // find the first buffer we will read from given the distance
-        read_history_buffer_buffer_idx = (history_buffer_buffer_idx - dist_small) & history_buffer_buffer_idx_mask;
-
-        // find the starting read index for each buffer, as well as compute
-        // the shuffle vector for shuffling the data to the output
-        #pragma unroll
-        for (int i = 0; i < literals_per_cycle; i++) {
-          // the buffer index
-          HistBufBufIdxT buf_idx = (read_history_buffer_buffer_idx + HistBufBufIdxT(i)) & history_buffer_buffer_idx_mask;
-
-          // compute the starting index for buffer 'buf_idx' (in range
-          // [0, literals_per_cycle))
-          HistBufIdxT starting_read_idx_for_this_buf = (history_buffer_idx[buf_idx] - ((dist - i) / literals_per_cycle)) - 1;
-          if (buf_idx == history_buffer_buffer_idx) {
-            starting_read_idx_for_this_buf += 1;
-          }
-          read_history_buffer_idx[buf_idx] = starting_read_idx_for_this_buf & history_buffer_idx_mask;
-
-          if (dist > i) {
-            read_history_shuffle_idx[i] = buf_idx;
-          } else {
-            // this special case happens whenever dist < literals_per_cycle
-            // and we need to repeat one of the earlier elements
-            // idx_back = dist_small + ((i - dist_small) % dist_small);
-            HistBufBufIdxT idx_back = mod_lut[dist_small - 1][i - 1];
-            read_history_shuffle_idx[i] = (history_buffer_buffer_idx - idx_back) & history_buffer_buffer_idx_mask;
-          }
-        }
-      }
-
-      if (reading_history) {
-        // grab from each of the history buffers
-        unsigned char historical_bytes[literals_per_cycle];
-
-        // using a tuple and static unroller because of these cases:
-        //    https://hsdes.intel.com/appstore/article/#/14015868135
-        //    https://hsdes.intel.com/appstore/article/#/14015867379
-        fpga_tools::UnrolledLoop<literals_per_cycle>([&](auto i) {
-          auto idx_in_buf = read_history_buffer_idx[i];
-          historical_bytes[i] = history_buffer.template get<i>()[idx_in_buf];
-
-          // also check the cache to see if it is there
-          #pragma unroll
-          for (int j = 0; j < kCacheDepth + 1; j++) {
-            if (history_buffer_cache_idx[i][j] == idx_in_buf) {
-              historical_bytes[i] = history_buffer_cache_val[i][j];
-            }
-          }
-        });
-
-        // shuffle the elements read from the history to the output
-        #pragma unroll
-        for (int i = 0; i < literals_per_cycle; i++) {
-          out_data.literal[i] = historical_bytes[read_history_shuffle_idx[i]];
-        }
-
-        if (history_counter < literals_per_cycle) {
-          out_data.valid_count = history_counter;
-        } else {
-          out_data.valid_count = literals_per_cycle;
-        }
-
-        // update the history read index
-        #pragma unroll
-        for (int i = 0; i < literals_per_cycle; i++) {
-          read_history_buffer_idx[i] = (read_history_buffer_idx[i] + ac_uint<1>(1)) & history_buffer_idx_mask;
-        }
-
-        // update whether we are still reading the history
-        reading_history = reading_history_next;
-        reading_history_next = history_counter > literals_per_cycle * 2;
-        history_counter -= literals_per_cycle;
-      }
-      
-      if (!done && data_valid) {
-        // compute the valid bitmap for the writes
-        bool write_bitmap[literals_per_cycle];
-        [[intel::fpga_register]] HistBufBufIdxT shuffle_vec[literals_per_cycle];
-        
-        #pragma unroll
-        for (int i = 0; i < literals_per_cycle; i++) {
-          HistBufBufIdxT buf_idx = (history_buffer_buffer_idx + i) & history_buffer_buffer_idx_mask;
-          write_bitmap[buf_idx] = i < out_data.valid_count;
-          shuffle_vec[buf_idx] = i;
-        }
-
-        // write to the history buffers what we are sending out
-        // using a tuple and static unroller because of these cases:
-        //    https://hsdes.intel.com/appstore/article/#/14015868135
-        //    https://hsdes.intel.com/appstore/article/#/14015867379
-        fpga_tools::UnrolledLoop<literals_per_cycle>([&](auto i) {
-          if (write_bitmap[i]) {
-            // grab the literal to write out
-            auto literal_out = out_data.literal[shuffle_vec[i]];
-
-            // the index into the buffer
-            HistBufIdxT idx_in_buf = history_buffer_idx[i];
-            history_buffer.template get<i>()[idx_in_buf] = literal_out;
-
-            // write new value into cache as well
-            history_buffer_cache_val[i][kCacheDepth] = literal_out;
-            history_buffer_cache_idx[i][kCacheDepth] = idx_in_buf;
-
-            // Cache is just a shift register, so shift the shift reg. Pushing
-            // into the back of the shift reg is done above.
-            #pragma unroll
-            for (int j = 0; j < kCacheDepth; j++) {
-              history_buffer_cache_val[i][j] = history_buffer_cache_val[i][j + 1];
-              history_buffer_cache_idx[i][j] = history_buffer_cache_idx[i][j + 1];
-            }
-
-            // update the history buffer index
-            history_buffer_idx[i] = (history_buffer_idx[i] + ac_uint<1>(1)) & history_buffer_idx_mask;
-          }
-        });
-
-        // update the most recent buffer
-        history_buffer_buffer_idx = (history_buffer_buffer_idx + out_data.valid_count) & history_buffer_buffer_idx_mask;
-
-        // write the output to the pipe
-        OutPipe::write(OutPipeBundleT(out_data));
-      }
-    } while (!done);
-    
-    OutPipe::write(OutPipeBundleT(true));
-  });
-}
-
-template<typename InPipe, typename OutPipe>
-event SubmitHuffmanDecoderKernel(queue& q) {
-  return q.single_task<HuffmanDecoderKernelID>([=] {
-    using OutPipeBundleT = FlagBundle<HuffmanData>;
-
-    BitStreamT bbs;
-    bool last_block;
-    bool done_reading = false;
-
-    [[intel::disable_loop_pipelining]]
-    do {
-      ac_uint<2> first_table_state = 0;
-      ac_uint<1> last_block_num;
-      ac_uint<2> type;
-      ac_uint<9> numlitlencodes;
-      ac_uint<6> numdistcodes;
-      ac_uint<5> numcodelencodes;
-      bool parsing_first_table;
-      unsigned short codelencodelen_count = 0;
-
-      // read in the first byte and add it to the byte bit stream
-      auto first_pipe_data = InPipe::read();
-      bbs.NewByte(first_pipe_data.data);
-
-      // read the first three bits
-      ac_uint<3> first_three_bits = bbs.ReadUInt<3>();
-      bbs.Shift(3);
-
-      // first bit indicates whether this is the last block
-      last_block = (first_three_bits.slc<1>(0) == 1);
-
-      // next 2 bits indicate the block type
-      //    0: uncompressed
-      //    1: static huffman
-      //    2: dynamic huffman
-      //    3: reserved
-      type = first_three_bits.slc<2>(1);
-      bool is_uncompressed_block = (type == 0);
-      bool is_static_huffman_block = (type == 1);
-      bool is_dynamic_huffman_block = (type == 2);
-
-      // if this is an uncompressed block, we must realign to a byte boundary
-      if (is_uncompressed_block) {
-        bbs.AlignToByteBoundary();
-      }
-
-      // only parse the first table for dynamically compressed blocks
-      parsing_first_table = is_dynamic_huffman_block;
-
-      constexpr unsigned short codelencodelen_idxs[] =
-        {16, 17, 18, 0, 8, 7, 9, 6, 10, 5, 11, 4, 12, 3, 13, 2, 14, 1, 15};
-
-      [[intel::fpga_register]] ac_uint<3> codelencodelen[19];
-      #pragma unroll
-      for (int i = 0; i < 19; i++) { codelencodelen[i] = 0; }
-
-      // NOTE: this loop is not the main processing loop and therefore is
-      // not critical (low trip count). However, the compiler doesn't know that
-      // and tries to optimize for throughput (~Fmax/II). However, we don't want
-      // this loop to be our Fmax bottleneck, so increase the II.
-      [[intel::initiation_interval(4)]]
-      while (parsing_first_table) {
-        if (bbs.HasSpaceForByte()) {
-          bool read_valid;
-          auto pd = InPipe::read(read_valid);
-
-          if (read_valid) {
-            unsigned char c = pd.data;
-            bbs.NewByte(c);
-          }
-        }
-
-        if (bbs.Size() >= 5) {
-          if (first_table_state == 0) {
-            numlitlencodes = bbs.ReadUInt(5) + ac_uint<9>(257);
-            bbs.Shift(5);
-            first_table_state = 1;
-          } else if (first_table_state == 1) {
-            numdistcodes = bbs.ReadUInt(5) + ac_uint<1>(1);
-            bbs.Shift(5);
-            first_table_state = 2;
-          } else if (first_table_state == 2) {
-            numcodelencodes = bbs.ReadUInt(4) + ac_uint<3>(4);
-            bbs.Shift(4);
-            first_table_state = 3;
-          } else if (codelencodelen_count < numcodelencodes) {
-            auto tmp = bbs.ReadUInt(3);
-            bbs.Shift(3);
-            codelencodelen[codelencodelen_idxs[codelencodelen_count]] = tmp;
-            codelencodelen_count++;
-            parsing_first_table = (codelencodelen_count != numcodelencodes);
-          }
-        }
-      }
-
-      [[intel::fpga_register]] ac_uint<8> codelencode_map_first_code[8];
-      [[intel::fpga_register]] ac_uint<8> codelencode_map_last_code[8];
-      [[intel::fpga_register]] ac_uint<5> codelencode_map_base_idx[8];
-      [[intel::fpga_register]] ac_uint<5> codelencode_map[19];
-
-      if (is_dynamic_huffman_block) {
-        ac_uint<8> codelencode_map_next_code = 0;
-        ac_uint<5> codelencode_map_counter = 0;
-        for (unsigned char codelen = 1; codelen <= 8; codelen++) {
-          codelencode_map_next_code <<= 1;
-          codelencode_map_first_code[codelen - 1] = codelencode_map_next_code;
-          codelencode_map_base_idx[codelen - 1] = codelencode_map_counter;
-          for (unsigned short symbol = 0; symbol < 19; symbol++) {
-            auto inner_codelen = codelencodelen[symbol];
-            if (inner_codelen == codelen) {
-              codelencode_map[codelencode_map_counter] = symbol;
-              codelencode_map_counter++;
-              codelencode_map_next_code++; 
-            }
-          }
-          codelencode_map_last_code[codelen - 1] = codelencode_map_next_code;
-        }
-      }
-
-      // length of codelens is MAX(numlitlencodes + numdistcodes)
-      // = MAX((2^5 + 257) + (2^5 + 1)) = 322
-      ac_uint<15> codelens[322];
-      ac_uint<9> total_codes_second_table = numlitlencodes + numdistcodes;
-      decltype(total_codes_second_table) codelens_idx = 0;
-      bool decoding_next_symbol = true;
-      ac_uint<8> runlen; // MAX = (2^7 + 11)
-      int onecount = 0, otherpositivecount = 0;
-      ac_uint<15> extend_symbol;
-
-      // static codelens ROM (for static huffman encoding)
-      constexpr int static_numlitlencodes = 288;
-      constexpr int static_numdistcodes = 32;
-      constexpr int static_totalcodes =
-          static_numlitlencodes + static_numdistcodes;
-      constexpr auto static_codelens = [] {
-        std::array<unsigned short, 320> a{};
-        // literal codes
-        for (int i = 0; i < static_numlitlencodes; i++) {
-          if (i < 144) {
-            a[i] = 8;
-          } else if (i < 144 + 112) {
-            a[i] = 9;
-          } else if (i < 144 + 112 + 24) {
-            a[i] = 7;
-          } else {
-            a[i] = 8;
-          }
-        }
-        // distance codes
-        for (int i = 0; i < static_numdistcodes; i++) {
-          a[static_numlitlencodes + 1] = 5;
-        }
-        return a;
-      }();
-
-      // for a static huffman block, the number of literal and distance codes
-      // is constant
-      if (is_static_huffman_block) {
-        numlitlencodes = static_numlitlencodes;
-        numdistcodes = static_numdistcodes;
-      }
-
-      // for static huffman block, initialize codelens with static codelens ROM
-      if (is_static_huffman_block) {
-        for (int i = 0; i < static_totalcodes; i++) {
-          codelens[i] = static_codelens[i];
-        }
-      }
-
-      // NOTE: this loop is not the main processing loop and therefore is
-      // not critical (low trip count). However, the compiler doesn't know that
-      // and tries to optimize for throughput (~Fmax/II). However, we don't want
-      // this loop to be our Fmax bottleneck, so increase the II.
-      [[intel::initiation_interval(5)]]
-      while ((codelens_idx < total_codes_second_table)
-             && is_dynamic_huffman_block) {
-        // read in another byte if we have space for it
-        if (bbs.HasSpaceForByte()) {
-          bool read_valid;
-          auto pd = InPipe::read(read_valid);
-
-          if (read_valid) {
-            unsigned char c = pd.data;
-            done_reading = pd.flag;
-            bbs.NewByte(c);
-          }
-        }
-
-        if (decoding_next_symbol) {
-          // decoding the next code symbol, so make sure we have enough bits to
-          // do so 15 bits is the maximum bits to read both a symbol and the
-          // extra run length bits (max 8 bits for the symbol, max 7 bits for
-          // extra run length)
-          if (bbs.Size() >= 15) {
-            // read 15 bits
-            ac_uint<15> next_bits = bbs.ReadUInt<15>();
-
-            // find all possible dynamic run lengths
-            // the symbol could be from 1 to 8 bits long and the number of extra
-            // bits to read for the run length could be either 2, 3, or 7 bits
-            // (3 possibilities in 'runlen_bits').
-            [[intel::fpga_register]] ac_uint<7> extra_bit_vals[8][3];
-            constexpr int runlen_bits[] = {2, 3, 7};
-            #pragma unroll
-            for (int out_codelen = 1; out_codelen <= 8; out_codelen++) {
-              #pragma unroll
-              for (int j = 0; j < 3; j++) {
-                ac_uint<7> codebits_tmp(0);
-                #pragma unroll
-                for (int bit = 0; bit < runlen_bits[j]; bit++) {
-                  codebits_tmp[bit] = next_bits[out_codelen + bit] & 0x1;
-                }
-                extra_bit_vals[out_codelen - 1][j] = codebits_tmp;
-              }
-            }
-
-            // decode all possible code symbols from 1 to 8 bits
-            ac_uint<8> codelencode_valid_bitmap(0);
-            ac_uint<5> codelencode_offset[8];
-            ac_uint<5> codelencode_base_idx[8];
-
-            #pragma unroll
-            for (int codelen = 1; codelen <= 8; codelen++) {
-              ac_uint<8> codebits_tmp(0);
-              #pragma unroll
-              for (int bit = 0; bit < codelen; bit++) {
-                codebits_tmp[codelen - bit - 1] = next_bits[bit] & 0x1;
-              }
-              unsigned char codebits = codebits_tmp;
-
-              auto base_idx = codelencode_map_base_idx[codelen - 1];
-              auto first_code = codelencode_map_first_code[codelen - 1];
-              auto last_code = codelencode_map_last_code[codelen - 1];
-              
-              codelencode_base_idx[codelen - 1] = base_idx;
-              codelencode_valid_bitmap[codelen - 1] =
-                  ((codebits >= first_code) && (codebits < last_code)) ? 1 : 0;
-              
-              codelencode_offset[codelen - 1] = codebits - first_code;
-            }
-
-            // find the shortest matching code symbol
-            ac_uint<3> shortest_match_len = ctz(codelencode_valid_bitmap) + 1;
-            ac_uint<5> base_idx = codelencode_base_idx[shortest_match_len - 1];
-            ac_uint<5> offset = codelencode_offset[shortest_match_len - 1];
-
-            // get the decoded symbol
-            auto symbol = codelencode_map[base_idx + offset];
-
-            // max shift amount will be 15 (8 bits for symbol, 7 for run length)
-            ac_uint<4> shift_amount;
-
-            // do logic based on symbol value
-            if (symbol <= 15) {
-              // ADD SYMBOL
-              codelens[codelens_idx++] = symbol;
-              decoding_next_symbol = true;
-              if (codelens_idx >= numlitlencodes) {
-                if (symbol == 1) {
-                  onecount++;
-                } else if (symbol > 0) {
-                  otherpositivecount++;
-                }
-              }
-              shift_amount = shortest_match_len;
-            } else if (symbol == 16) {
-              // READ 2-BIT RUN LENGTH, ADD 3, AND EXTEND LAST ELEMENT
-              runlen = extra_bit_vals[shortest_match_len - 1][0] + 3;
-              decoding_next_symbol = false;
-              extend_symbol = codelens[codelens_idx-1];
-              shift_amount = shortest_match_len + 2;
-            } else if (symbol == 17) {
-              // READ 3-BIT RUN LENGTH, ADD 3, AND EXTEND WITH 0's
-              runlen = extra_bit_vals[shortest_match_len - 1][1] + 3;
-              decoding_next_symbol = false;
-              extend_symbol = 0;
-              shift_amount = shortest_match_len + 3;
-            } else if (symbol == 18) {
-              // READ 7-BIT RUN LENGTH, ADD 11, AND EXTEND WITH 0's
-              runlen = extra_bit_vals[shortest_match_len - 1][2] + 11;
-              decoding_next_symbol = false;
-              extend_symbol = 0;
-              shift_amount = shortest_match_len + 7;
-            }
-
-            // shift the bit stream
-            bbs.Shift(shift_amount);
-          }
-        } else {
-          // extending codelens
-          codelens[codelens_idx++] = extend_symbol;
-          if (codelens_idx >= numlitlencodes) {
-            if (extend_symbol == 1) {
-              onecount++;
-            } else if (extend_symbol > 0) {
-              otherpositivecount++;
-            }
-          }
-
-          // decrement the run length
-          runlen--;
-
-          // start reading decoding symbols again when runlen == 0
-          decoding_next_symbol = (runlen == 0);
-        }
-      }
-
-      // handle the case where only one distance code is defined add a dummy
-      // invalid code to make the Huffman tree complete
-      if (onecount == 1 && otherpositivecount == 0 && is_dynamic_huffman_block) {
-        int extend_amount = 32 - numdistcodes;
-        for (int i = 0; i < extend_amount; i++) {
-          codelens[numlitlencodes + numdistcodes + i] = 0;
-        }
-        codelens[numlitlencodes + 31] = 1;
-        numdistcodes += extend_amount;
-      }
-
-      // the first table is decoded, so now it is time to decode the second
-      // table, which is actually two tables:
-      //  literal table (symbols and lengths for the {length, distance} pair)
-      //  distance table (the distances for the {length, distance} pair) 
-      [[intel::fpga_register]] ac_uint<15> lit_map_first_code[15];
-      [[intel::fpga_register]] ac_uint<15> lit_map_last_code[15];
-      [[intel::fpga_register]] ac_uint<9> lit_map_base_idx[15];
-      [[intel::fpga_register]] ac_uint<9> lit_map[286];
-
-      [[intel::fpga_register]] ac_uint<15> dist_map_first_code[15];
-      [[intel::fpga_register]] ac_uint<15> dist_map_last_code[15];
-      [[intel::fpga_register]] ac_uint<5> dist_map_base_idx[15];
-      [[intel::fpga_register]] ac_uint<5> dist_map[32];
-
-      ac_uint<15> lit_map_next_code = 0;
-      ac_uint<9> lit_map_counter = 0;
-      for (unsigned char codelen = 1; codelen <= 15; codelen++) {
-        lit_map_next_code <<= 1;
-        lit_map_first_code[codelen - 1] = lit_map_next_code;
-        lit_map_base_idx[codelen - 1] = lit_map_counter;
-        for (unsigned short symbol = 0; symbol < numlitlencodes; symbol++) {
-          auto inner_codelen = codelens[symbol];
-          if (inner_codelen == codelen) {
-            lit_map[lit_map_counter] = symbol;
-            lit_map_counter++;
-            lit_map_next_code++; 
-          }
-        }
-        lit_map_last_code[codelen - 1] = lit_map_next_code;
-      }
-
-      ac_uint<15> dist_map_next_code = 0;
-      ac_uint<5> dist_map_counter = 0;
-      for (unsigned char codelen = 1; codelen <= 15; codelen++) {
-        dist_map_next_code <<= 1;
-        dist_map_first_code[codelen - 1] = dist_map_next_code;
-        dist_map_base_idx[codelen - 1] = dist_map_counter;
-        for (unsigned short symbol = 0; symbol < numdistcodes; symbol++) {
-          auto inner_codelen = codelens[numlitlencodes + symbol];
-          if (inner_codelen == codelen) {
-            dist_map[dist_map_counter] = symbol;
-            dist_map_counter++;
-            dist_map_next_code++; 
-          }
-        }
-        dist_map_last_code[codelen - 1] = dist_map_next_code;
-      }
-
-      // indicates whether we are reading a distance (or literal) currently
-      bool reading_distance = false;
-
-      // true is the stop code (256) has been decoded and the block is done
-      bool stop_code_hit = false;
-
-      // true when output is ready. the output will be either a character or
-      // a length distance pair (see HuffmanData struct)
-      bool out_ready = false;
-
-      // the output data which is either a character or a length distance pair
-      HuffmanData out_data;
-
-      ac_uint<9> lit_symbol;
-      ac_uint<5> dist_symbol;
-
-      // main processing loop
-      // the II of this main loop can be controlled from the command line using
-      // the -DHUFFMAN_MAIN_LOOP_II=<desired II>. By default, we let the
-      // the compiler choose the Fmax/II to maximize throughput
-#ifdef HUFFMAN_MAIN_LOOP_II
-      [[intel::initiation_interval(HUFFMAN_MAIN_LOOP_II)]]
-#endif
-      do {
-        // read in new data if the ByteBitStream has space for it and we aren't
-        // done reading from the input pipe
-        if (bbs.HasSpaceForByte()) {
-          bool read_valid;
-          auto pd = InPipe::read(read_valid);
-
-          if (read_valid) {
-            unsigned char c = pd.data;
-            done_reading = pd.flag;
-            bbs.NewByte(c);
-          }
-        }
-        
-        if (is_uncompressed_block) {
-          // for uncompressed blocks, simply read an 8-bit character from the
-          // stream and write it to the output
-          if (bbs.Size() >= 8) {
-            ac_uint<8> byte = bbs.ReadUInt<8>();
-            out_data.len_or_sym = byte;
-            out_data.dist_or_flag = -1;
-            out_ready = true;
-          }
-        } else if (bbs.Size() >= 30) {
-          // read the next 30 bits (we know we have them)
-          ac_uint<30> next_bits = bbs.ReadUInt<30>();
-
-          // find all possible dynamic lengths
-          [[intel::fpga_register]] ac_uint<5> lit_extra_bit_vals[15][5];
-          #pragma unroll
-          for (int out_codelen = 1; out_codelen <= 15; out_codelen++) {
-            #pragma unroll
-            for (int in_codelen = 1; in_codelen <= 5; in_codelen++) {
-              ac_uint<5> codebits_tmp(0);
-              #pragma unroll
-              for (int bit = 0; bit < in_codelen; bit++) {
-                codebits_tmp[bit] = next_bits[out_codelen + bit] & 0x1;
-              }
-              lit_extra_bit_vals[out_codelen - 1][in_codelen - 1] = codebits_tmp;
-            }
-          }
-
-          // find all possible dynamic distances
-          [[intel::fpga_register]] ac_uint<15> dist_extra_bit_vals[15][15];
-          #pragma unroll
-          for (int out_codelen = 1; out_codelen <= 15; out_codelen++) {
-            #pragma unroll
-            for (int in_codelen = 1; in_codelen <= 15; in_codelen++) {
-              ac_uint<15> codebits_tmp(0);
-              #pragma unroll
-              for (int bit = 0; bit < in_codelen; bit++) {
-                codebits_tmp[bit] = next_bits[out_codelen + bit] & 0x1;
-              }
-              dist_extra_bit_vals[out_codelen - 1][in_codelen - 1] = codebits_tmp;
-            }
-          }
-
-          // find all possible code lengths and offsets
-          ac_uint<15> lit_codelen_valid_bitmap(0), dist_codelen_valid_bitmap(0);
-          ac_uint<9> lit_codelen_offset[15], lit_codelen_base_idx[15];
-          ac_uint<5> dist_codelen_offset[15], dist_codelen_base_idx[15];
-
-          #pragma unroll
-          for (unsigned char codelen = 1; codelen <= 15; codelen++) {
-            // NOTE: could use fpga_tools::UnrolledLoop so that we don't need to
-            // use the maximum codelen.
-            // Blocking cases:
-            //    https://hsdes.intel.com/appstore/article/#/14015701865
-            //    https://hsdes.intel.com/appstore/article/#/14015701865
-            ac_uint<15> codebits(0);
-            #pragma unroll
-            for (unsigned char bit = 0; bit < codelen; bit++) {
-              codebits[codelen - bit - 1] = next_bits[bit];
-            }
-
-            auto lit_base_idx = lit_map_base_idx[codelen - 1];
-            auto lit_first_code = lit_map_first_code[codelen - 1];
-            auto lit_last_code = lit_map_last_code[codelen - 1];
-            auto dist_base_idx = dist_map_base_idx[codelen - 1];
-            auto dist_first_code = dist_map_first_code[codelen - 1];
-            auto dist_last_code = dist_map_last_code[codelen - 1];
-
-            lit_codelen_valid_bitmap[codelen - 1]
-                = ((codebits >= lit_first_code) &&
-                   (codebits < lit_last_code)) ? 1 : 0; 
-            lit_codelen_base_idx[codelen - 1] = lit_base_idx;
-            lit_codelen_offset[codelen - 1] = codebits - lit_first_code;
-
-            dist_codelen_valid_bitmap[codelen - 1] =
-                ((codebits >= dist_first_code) &&
-                 (codebits < dist_last_code)) ? 1 : 0;
-            dist_codelen_base_idx[codelen - 1] = dist_base_idx;
-            dist_codelen_offset[codelen - 1] = codebits - dist_first_code;
-          }
-
-          // find the shortest matching length, which is the next decoded symbol
-          auto lit_shortest_match_len = ctz(lit_codelen_valid_bitmap) + 1;
-          auto dist_shortest_match_len = ctz(dist_codelen_valid_bitmap) + 1;
-
-          // get the base index and offset based on the shortest match length
-          auto lit_base_idx = lit_codelen_base_idx[lit_shortest_match_len - 1];
-          auto lit_offset = lit_codelen_offset[lit_shortest_match_len - 1];
-          auto dist_base_idx = dist_codelen_base_idx[dist_shortest_match_len - 1];
-          auto dist_offset = dist_codelen_offset[dist_shortest_match_len - 1];
-          ac_uint<9> lit_idx = lit_base_idx + lit_offset;
-          ac_uint<9> dist_idx = dist_base_idx + dist_offset;
-
-          // lookup the symbol using base_idx and offset
-          lit_symbol = lit_map[lit_idx];
-          dist_symbol =  dist_map[dist_idx];
-
-          // we will either shift by shortest_match_len or by
-          // shortest_match_len + num_extra_bits based on whether we read a
-          // length, distance and/or its extra bits.
-          // maximum value for shift_amount = 15 + 15 = 30
-          ac_uint<5> shift_amount;
-
-          if (!reading_distance) {
-            shift_amount = lit_shortest_match_len;
-            // currently parsing a symbol or length (same table)
-            if (lit_symbol == 256) {
-              // stop code hit, done this block
-              stop_code_hit = true;
-              out_ready = false;
-            } else if (lit_symbol < 256) {
-              // decoded a regular character
-              out_data.len_or_sym = lit_symbol;
-              out_data.dist_or_flag = -1;
-              out_ready = true;
-            } else if (lit_symbol <= 264) {
-              // decoded a length with a static value
-              out_data.len_or_sym = lit_symbol - ac_uint<9>(254);
-              reading_distance = true;
-            } else if (lit_symbol <= 284) {
-              // decoded a length with a dynamic value
-              ac_int<5> lit_symbol_small = lit_symbol.template slc<5>(0);
-              // (lit_symbol - 261) / 4
-              ac_uint<3> num_extra_bits = (lit_symbol_small - ac_uint<5>(5)) >> 2;  
-              auto extra_bits_val = lit_extra_bit_vals[lit_shortest_match_len - 1][num_extra_bits - 1];
-              // ((((lit_symbol - 265) % 4) + 4) << num_extra_bits) + ac_uint<2>(3) + extra_bits_val
-              out_data.len_or_sym = ((((lit_symbol_small - ac_uint<5>(9)) & 0x3) + ac_uint<3>(4)) << num_extra_bits) + ac_uint<2>(3) + extra_bits_val;
-              shift_amount = lit_shortest_match_len + num_extra_bits;
-              reading_distance = true;
-            } else if (lit_symbol == 285) {
-              // decoded a length with a static value
-              out_data.len_or_sym = 258;
-              reading_distance = true;
-            } // else error, ignored
-          } else {
-            shift_amount = dist_shortest_match_len;
-            // currently decoding a distance symbol
-            if (dist_symbol <= 3) {
-              // decoded a distance with a static value
-              out_data.dist_or_flag = dist_symbol + 1;
-            } else {
-              // decoded a distance with a dynamic value
-              // NOTE: should be <= 29, but not doing error checking
-              ac_uint<4> num_extra_bits = (dist_symbol >> 1) - ac_uint<1>(1);
-              auto extra_bits_val = dist_extra_bit_vals[dist_shortest_match_len - 1][num_extra_bits - 1];
-              out_data.dist_or_flag = (((dist_symbol & 0x1) + ac_uint<2>(2)) << num_extra_bits) + ac_uint<1>(1) + extra_bits_val;
-              shift_amount = dist_shortest_match_len + num_extra_bits;
-            }
-            out_ready = true;
-            reading_distance = false;
-          }
-
-          // shift based on how many bits we read
-          bbs.Shift(shift_amount);
-        }
-
-        // output data to downstream kernel if ready
-        if (out_ready) {
-          OutPipe::write(OutPipeBundleT(out_data));
-          out_ready = false;
-        }
-      } while (!stop_code_hit);
-    } while (!last_block);
-
-    // notify the downstream kernel that we are done
-    OutPipe::write(OutPipeBundleT(true));
-
-    // read out the remaining data from the pipe
-    // NOTE: don't really care about performance here
-    while (!done_reading) {
-      bool read_valid;
-      auto pd = InPipe::read(read_valid);
-      done_reading = pd.flag && read_valid;
-    }
-  });
-}
-
-template<typename InPipe, typename OutPipe>
-event SubmitHeaderKernel(queue& q, int in_count, HeaderData* hdr_data_ptr, int* crc_ptr, int* out_count_ptr) {
-  return q.single_task<HeaderKernelID>([=]() [[intel::kernel_args_restrict]] {
-    using OutPipeBundleT = FlagBundle<unsigned char>;
-    // read magic number (2 bytes)
-    // read compression method (1 byte)
-    // read 'flags' (1 byte)
-    // read time (4 bytes)
-    // read extra flags (1 byte)
-    // read OS (1 byte)
-    // read based on flags:
-    //    if flags & 0x01 != 0: Flag = Text
-    //    if flags & 0x04 != 0: Flag = Errata, read 2 bytes for 'length', read 'length' more bytes
-    //    if flags & 0x08 != 0: Filename, read nullterminated string
-    //    if flags & 0x02 != 0: CRC-16, read 2 bytes
-    //    if flags & 0x10 != 0: Comment, read nullterminated string
-    int i = 0;
-    bool i_in_range = 0 < in_count;
-    bool i_next_in_range = 1 < in_count;
-    short state_counter = 0;
-    short errata_len = 0;
-    unsigned char curr_byte;
-    HeaderState state = MagicNumber;
-    device_ptr<HeaderData> hdr_data(hdr_data_ptr);
-
-    unsigned char header_magic[2];
-    unsigned char header_compression_method;
-    unsigned char header_flags;
-    unsigned char header_time[4];
-    unsigned char header_os;
-    unsigned char header_filename[256];
-    unsigned char header_crc[2];
-    header_filename[0] = '\0';
-
-    device_ptr<int> crc(crc_ptr);
-    device_ptr<int> out_count(out_count_ptr);
-
-    // NOTE: this loop is not the main processing loop and therefore is
-    // not critical (low trip count). However, the compiler doesn't know that
-    // and tries to optimize for throughput (~Fmax/II). However, we don't want
-    // this loop to be our Fmax bottleneck, so increase the II.
-    [[intel::initiation_interval(4)]]
-    while (state != SteadyState) {
-      curr_byte = InPipe::read();
-
-      switch (state) {
-        case MagicNumber: {
-          header_magic[state_counter] = curr_byte;
-          state_counter++;
-          if (state_counter == 2) {
-            state = CompressionMethod;
-            state_counter = 0;
-          }
-          break;
-        }
-        case CompressionMethod: {
-          header_compression_method = curr_byte;
-          state = Flags;
-          break;
-        }
-        case Flags: {
-          header_flags = curr_byte;
-          state = Time;
-          break;
-        }
-        case Time: {
-          header_time[state_counter] = curr_byte;
-          state_counter++;
-          if (state_counter == 4) {
-            state = ExtraFlags;
-            state_counter = 0;
-          }
-          break;
-        }
-        case ExtraFlags: {
-          state = OS;
-          break;
-        }
-        case OS: {
-          header_os = curr_byte;
-          if (header_flags & 0x04) {
-            state = Errata;
-          } else if(header_flags & 0x08) {
-            state = Filename;
-          } else if(header_flags & 0x02) {
-            state = CRC;
-          } else if(header_flags & 0x10) {
-            state = Comment;
-          } else {
-            state = SteadyState;
-          }
-          break;
-        }
-        case Errata: {
-          if (state_counter == 0) {
-            errata_len |= curr_byte;
-            state_counter++;
-          } else if (state_counter == 1) {
-            errata_len |= (curr_byte << 8);
-            state_counter++;
-          } else {
-            if ((state_counter - 2) == errata_len) {
-              if(header_flags & 0x08) {
-                state = Filename;
-              } else if(header_flags & 0x02) {
-                state = CRC;
-              } else if(header_flags & 0x10) {
-                state = Comment;
-              } else {
-                state = SteadyState;
-              }
-              state_counter = 0;
-            } else {
-              state_counter++;
-            }
-          }
-          break;
-        }
-        case Filename: {
-          header_filename[state_counter] = curr_byte;
-          if (curr_byte == '\0') {
-            if(header_flags & 0x02) {
-              state = CRC;
-            } else if(header_flags & 0x10) {
-              state = Comment;
-            } else {
-              state = SteadyState;
-            }
-            state_counter = 0;
-          } else {
-            state_counter++;
-          }
-          break;
-        }
-        case CRC: {
-          if (state_counter == 0) {
-            header_crc[0] = curr_byte;
-            state_counter++;
-          } else if (state_counter == 1) {
-            header_crc[1] = curr_byte;
-            state_counter++;
-          } else {
-            if(header_flags & 0x10) {
-              state = Comment;
-            } else {
-              state = SteadyState;
-            }
-            state_counter = 0;
-          }
-          break;
-        }
-        case Comment: {
-          if (curr_byte == '\0') {
-            state = SteadyState;
-            state_counter = 0;
-          } else {
-            state_counter++;
-          }
-          break;
-        }
-        default: {
-          break;
-        }
-      }
-
-      i_in_range = i_next_in_range;
-      i_next_in_range = i < (in_count - 2);
-      i++;
-    }
-
-    unsigned char crc_bytes[4];
-    unsigned char size_bytes[4];
-
-    while (i_in_range) {
-      bool valid_pipe_read;
-      curr_byte = InPipe::read(valid_pipe_read);
-
-      if (valid_pipe_read) {
-        int tmp = (in_count - i - 1);
-        if (tmp < 8) {
-          if (tmp < 4) {
-            size_bytes[3 - tmp] = curr_byte;
-          } else {
-            crc_bytes[7 - tmp] = curr_byte;
-          }
-        }
-        OutPipe::write(OutPipeBundleT(curr_byte, (i == (in_count-1))));
-
-        i_in_range = i_next_in_range;
-        i_next_in_range = i < (in_count - 2);
-        i++;
-      }
-    }
-
-    unsigned int crc_local = 0, size_local = 0;
-    for (int i = 0; i < 4; i++) {
-      crc_local |= (unsigned int)(crc_bytes[i]) << (i*8);
-      size_local |= (unsigned int)(size_bytes[i]) << (i*8);
-    }
-
-    // construct header data
-    HeaderData hdr_data_loc;
-    hdr_data_loc.magic[0] = header_magic[0];
-    hdr_data_loc.magic[1] = header_magic[1];
-    hdr_data_loc.compression_method = header_compression_method;
-    hdr_data_loc.flags = header_flags;
-    for(int i = 0; i < 4; i++) hdr_data_loc.time[i] = header_time[i];
-    hdr_data_loc.os = header_os;
-    for(int i = 0; i < 256; i++) hdr_data_loc.filename[i] = header_filename[i];
-    hdr_data_loc.crc[0] = header_crc[0];
-    hdr_data_loc.crc[1] = header_crc[1];
-
-    // write back header data, crc, and size
-    *hdr_data = hdr_data_loc;
-    *crc = crc_local;
-    *out_count = size_local;
-  });
-}
-
-template<typename InPipe, typename OutPipe, unsigned literals_per_cycle>
-std::vector<event> SubmitDecompressKernels(queue& q, int in_count,
-                                           HeaderData *hdr_data_out,
-                                           int *crc_out, int *size_out) {
-  using HeaderToHuffmanPipe =
-    ext::intel::pipe<HeaderToHuffmanPipeID, FlagBundle<unsigned char>>;
-  using HuffmanToLZ77Pipe =
-    ext::intel::pipe<HuffmanToLZ77PipeID, FlagBundle<HuffmanData>, 16>;  // TODO: experiment with depth
-  using LZ77ToLiteralStackerPipe =
-    ext::intel::pipe<LZ77ToLiteralStackerPipeID, FlagBundle<LiteralPack<literals_per_cycle>>>;
-
-  auto header_event = SubmitHeaderKernel<InPipe, HeaderToHuffmanPipe>(q, in_count, hdr_data_out, crc_out, size_out);
-  auto huffman_event = SubmitHuffmanDecoderKernel<HeaderToHuffmanPipe, HuffmanToLZ77Pipe>(q);
-  auto lz77_event = SubmitLZ77DecoderKernel<HuffmanToLZ77Pipe, LZ77ToLiteralStackerPipe, literals_per_cycle>(q);
-  auto lit_stacker_event = SubmitLiteralStackerKernel<LZ77ToLiteralStackerPipe, OutPipe, literals_per_cycle>(q);
-
-  return {header_event, huffman_event, lz77_event, lit_stacker_event};
-}
-
-#endif /* __DECOMPRESSOR_HPP__ */
